Trade interface

+++ Oregon Trail/src/Wagon.java
@@ -14,9 +14,9 @@ public class Wagon {
 	
 	// 	private ArrayList<MemeberHealth> wagon = new ArrayList<>();
 	private ArrayList<Equipment> wagon = new ArrayList<>();
-	private int foodWeight = 0;
-	private int totalWeight = 0;
-	private int consumableWeight = 0;
+	private float foodWeight = 0;
+	private float totalWeight = 0;
+	private float consumableWeight = 0;
 	
 	/**
 	 * Creates an Wagon object
@@ -31,7 +31,7 @@ public class Wagon {
 	public void addItem(Equipment item) {
 		// add the item to the array and get its weight
 		int weight = item.getWeight();
-		int quantity = item.getQuantity();
+		float quantity = item.getQuantity();
 		wagon.add(item);
 		
 		// check if the item is food/consumable and add that weight to the respective totals
@@ -50,10 +50,9 @@ public class Wagon {
 	 * the weight gets removed from the food weight as well
 	 * @param item - the item being removed from the array
 	 */
-	public void removeItem(Equipment item, int quantity) {
+	public void removeItem(Equipment item, float quantity) {
 		// remove the item from the array and get its weight
 		int weight = item.getWeight();
-		// wagon.remove(item);
 		
 		// check if the item is food/consumable and remove that weight from the respective totals
 		if (item instanceof Food) {
@@ -92,7 +91,7 @@ public class Wagon {
 	 * returns the total weight of the wagon
 	 * @return totalWeight - the total weight of the wagon
 	 */
-	public int getTotalWeight() {
+	public float getTotalWeight() {
 		return totalWeight;
 	}
 	
@@ -100,7 +99,7 @@ public class Wagon {
 	 * returns the food weight of the wagon
 	 * @return foodWeight - the weight of the food in the wagon
 	 */
-	public int getConsumableWeight() {
+	public float getConsumableWeight() {
 		return consumableWeight;
 	}
 }
